Always use assert.fail
As mentioned in sinon document, all failed assert calls
sinon.assert.fail().

The problem is, sinon try to integrate with other assert library, but
chai.assert.fail() has different argument signature:

  assert.fail = function (actual, expected, message, operator)

sinon thought message is the first argument of fail(), but it is not in
chai, cause message not displayed in mocha test runner.